Add php 8.2 to test pipeline
* Remove composer.lock from version control.
There were issues with peer dependencies in other libraries,
e.g. marein/symfony-lock-doctrine-migrations-bundle.
These issues don't exist here, but for consistency reasons,
the lockfile is also deleted in this library.
* Pin phpstan and phpcs to specific versions.
The reason is that they might break the pipeline when we download
the latest minor release, as they might adjust their rulesets.
This was the main reason for comitting lockfile in the past.